foreign key in Entry model

+++ server/controllers/entryController.js
@@ -9,6 +9,7 @@ class entryController {
     const time = new Date().toLocaleTimeString();
     const entry = req.body;
     entry.id = Entry.length + 1;
+    entry.userId = parseInt(req.user.id);
 
     const validateEntry = entrySchema.validate({
       id: entry.id, title: entry.title, description: entry.description,
@@ -22,66 +23,63 @@ class entryController {
     return res.status(201).json({
       status: 201,
       message: 'Entry successfully created',
-      data,
+      data: {
+        id: data.id,
+        title: data.title,
+        description: data.description,
+        createdOn: data.createdOn,
+      },
     });
   }
 
   static allEntries(req, res) {
-    return res.status(200).json({
-      status: 200,
-      data: Entry,
-    });
+    const myentries = Entry.filter(entry => entry.userId === parseInt(req.user.id));
+    return res.status(200).json({ status: 200, data: myentries });
   }
 
   static getEntry(req, res) {
     const id = parseInt(req.params.entryId);
     if (!Number.isInteger(id)) {
-      return res.status(400).json({
-        status: 400,
-        error: 'Please provide a valid id',
-      });
+      return res.status(400).json({ status: 400, error: 'Please provide a valid id' });
     }
     const entry = helperFunction.findById(id);
     if (!entry) {
-      return res.status(404).json({
-        status: 404,
-        error: 'Entry not found',
-      });
+      return res.status(404).json({ status: 404, error: 'Entry not found' });
     }
-    return res.status(200).json({
-      status: 200,
-      data: entry,
-    });
+    return res.status(200).json({ status: 200, data: entry });
   }
 
   static modifyEntry(req, res) {
     const id = parseInt(req.params.entryId);
     if (!Number.isInteger(id)) {
-      return res.status(400).json({
-        status: 400,
-        error: 'Please provide a valid id',
-      });
+      return res.status(400).json({ status: 400, error: 'Please provide a valid id' });
     }
     const entryIndex = helperFunction.findEntryIndex(id);
+
     if (!Entry[entryIndex]) {
-      return res.status(404).json({
-        status: 404,
-        error: 'Entry not found',
+      return res.status(404).json({ status: 404, error: 'Entry not found' });
+    }
+    if (Entry[entryIndex].userId === parseInt(req.user.id)) {
+      const validateEntry = entrySchema.validate({
+        id: Entry[entryIndex].id, title: req.body.title, description: req.body.description,
+      });
+      if (validateEntry.error) {
+        return res.status(400).json({ status: 400, error: validateEntry.error.details[0].message });
+      }
+      Entry[entryIndex].title = req.body.title;
+      Entry[entryIndex].description = req.body.description;
+      return res.status(200).json({
+        status: 200,
+        message: 'Entry successfully edited!',
+        data: {
+          id: Entry[entryIndex].id,
+          title: Entry[entryIndex].title,
+          description: Entry[entryIndex].description,
+          createdOn: Entry[entryIndex].createdOn,
+        },
       });
     }
-    const validateEntry = entrySchema.validate({
-      id: Entry[entryIndex].id, title: req.body.title, description: req.body.description,
-    });
-    if (validateEntry.error) {
-      return res.status(400).json({ status: 400, error: validateEntry.error.details[0].message });
-    }
-    Entry[entryIndex].title = req.body.title;
-    Entry[entryIndex].description = req.body.description;
-    return res.status(200).json({
-      status: 200,
-      message: 'Entry successfully edited!',
-      data: Entry[entryIndex],
-    });
+    return res.status(403).json({ status: 403, error: 'Operation forbiden' });
   }
 
   static deleteEntry(req, res) {
@@ -99,11 +97,14 @@ class entryController {
         error: 'Entry not found',
       });
     }
-    helperFunction.deleteEntryById(id);
-    return res.status(200).json({
-      status: 204,
-      message: 'Entry successfully deleted!',
-    });
+    if (entry.userId === parseInt(req.user.id)) {
+      helperFunction.deleteEntryById(id);
+      return res.status(200).json({
+        status: 204,
+        message: 'Entry successfully deleted!',
+      });
+    }
+    return res.status(403).json({ status: 403, error: 'Operation forbiden' });
   }
 }
